requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pytest==7.4.2
python-dateutil==2.8.2
schedule==1.2.0 